Metodos de interpolacion - trazador cubico/requirements.txt
numpy
pillow
matplotlib
opencv-python
scipy